Rename packages
add LinearMotion

+++ src/main/java/io/patriot_framework/generator/eventSimulator/Time/ContinuousTimeMinutes.java
@@ -31,7 +31,7 @@ public class ContinuousTimeMinutes extends AbstractContinuousTime {
         super();
     }
 
-    public ContinuousTimeMinutes(long timeInUnits) {
+    public ContinuousTimeMinutes(double timeInUnits) {
         super(timeInUnits);
     }
